Suppress MSBuild warning about non-assembly native code in .nupkgs
WixToolset.CoreNative includes the wixnative.exe in the lib folder to
place the executable as needed for tests. Since it is not a managed
assembly, NuGet complains that it is in the lib folder. We know what
we're doing (and we're not thrilled we have to do it) so suppress this
warning that shows up in many, many repos.